+++ tests/test_data_sources.py
@@ -12,9 +12,11 @@ def test_data_sources_import():
 
 
 @pytest.mark.skipif(os.getenv("DATA_SDK") != "yes", reason="Extra dependency required")
-def test_data_sources_import_cant_find(): # noqa
+def test_data_sources_import_cant_find():  # noqa
     try:
         from domino.data_sources import NotADatasourceConfig
+
+        assert NotADatasourceConfig
     except ImportError as exc:
         assert (
             "cannot import name 'NotADatasourceConfig' from 'domino.data_sources'"