+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to the `python-domino` library will be documented in this fi
 ### Changed
 * updated airflow dependency 2.2.4
 * Updated runs_stdout output
+* Fixed issues with import from domino_data
 
 ## 1.0.8