+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the `python-domino` library will be documented in this fi
 * Added black and flake8 formatting
 * Implemented pre-commit for black / flake8 formatting.
 * Added / updated dataset endpoint usability.
+* Added option to bypass certification verification
 
 ### Changed
 * updated airflow dependency 2.2.4