+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to the `python-domino` library will be documented in this fi
 * Fixed issues with import from domino_data
 * Updated nbconvert dependency version to 6.3.0
 * updated request user-agent to python-domino/{version}
+* Changed file handling for `files_upload`
 
 ## 1.0.8