+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the `python-domino` library will be documented in this fi
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+## 1.1.0
+
+### Added
 * Added black and flake8 formatting
 * Implemented pre-commit for black / flake8 formatting.
 * Added / updated dataset endpoint usability.